# utils/swagger.yaml
openapi :  '3.0.0'
info : 
  title :  Jungle Meet Apis Documentation
  version :  '1.0.0'
  description :  This is the API documentation for JungleMeet Queuing
servers :
  - url :  http://localhost:3000/v1  
    description :  Local Server
components:
  schemas:
...
    Post:
      type: object
      required:
        - title
        - content
        - bgImg
        - author
      properties:
        _id:
          type: string
          description: The Auto-generated id of a post
          example: 63338ae4d11e553b55138c9e
        title:
          type: string
          description: title of the post
          example: abd
        content:
          type: string
          description: content of the post
          example: 1234
        bgImg:
          type: string
          description: background image of the the post
          example: xxxx
        hashtag:
          type: string
          description: tag post
          example: horror
        author:
          type: string
          description: author of the post
          example: 6332d81195b55eda2a612059
        createdTime:
          type: string
          description: the time of creating the post
          example: 2022-10-03T01:33:24.275Z
        updatedTime:
          type: string
          description: the time of updating the post
          example: 2022-09-30T06:01:26.183Z
        visible:
          type: boolean
          description: status of deleting the post
          example: true
        resourceId:
          type: string
          description: resource id of the post
          example: xxxxxx
        viewCount:
          type: number
          description: the number of the post views
          example: 11
        postType:
          type: string
          description: the type of the post
          example: usePost
        follower:
          type: array
          items:
            type: string
          description: follower of the user
          example: [6332d81195b55eda2a612058, 6332d81195b55eda2a612057]
        like:
          type: array
          items:
            type: string
          description: the userId of users who like post
          example: [6332d81195b55eda2a612058, 6332d81195b55eda2a612057]
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
paths :
...
    /posts:
      get:
        tags:
          - post
        summary: get all posts
        description: get all posts
        operationId: getPosts
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Post'
          '404':
            description: Not found
    /posts/post:
      post:
        tags:
          - post
        summary: Add a new post
        description: Return created post
        operationId: createPost
        security:
          - cookieAuth: []
        requestBody:
          description:
            Add a new post
          content:
            application/json:
              schema:
                type: object
                required:
                  - title
                  - content
                properties:
                  title:
                    type: string
                    description: title of the post
                    example: abd
                  content:
                    type: string
                    description: content of the post
                    example: 1234
                  bgImg:
                    type: string
                    description: background image of the the post
                    example: xxxx
                  hashtag:
                    type: string
                    description: tag post
                    example: horror

        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Post'
          '404':
            description: Not Found
          '400':
            description: Title and content cannot be empty!
    /posts/{postId}:
      patch:
        tags:
          - post
        summary: Patch a post by id
        description: Patch a post
        operationId: patchPost
        security:
          - cookieAuth: []
        parameters:
          - name: postId
            required: true
            in: path
            description: the id of the post
            schema:
              type: string
        requestBody:
          description:
            patch a post by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                    description: title of the post
                    example: abd
                  content:
                    type: string
                    description: content of the post
                    example: 1234
                  bgImg:
                    type: string
                    description: background image of the the post
                    example: xxxx
                  hashtag:
                    type: string
                    description: tag post
                    example: horror

        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Post'
          '401':
            description: Only author can update post!
          '400':
            description: Title and content cannot be empty!
          '404':
            description: Not found
      put:
        tags:
          - post
        summary: Update an existing post
        description: Update an existing post by Id
        operationId: updatePost
        security:
          - cookieAuth: []
        parameters:
          - name: postId
            in: path
            description: ID of post to return
            schema:
              type: string
            required: true
        requestBody:
          description:
            Update an existent post on the forum
          content:
            application/json:
              schema:
                type: object
                required:
                  - title
                  - content
                properties:
                  title:
                    type: string
                    description: title of the post
                    example: Matrix
                  content:
                    type: string
                    description: content of the post
                    example: Great movie!
                  bgImg:
                    type: string
                    description: background image of the the post
                    example: xxxx
                  hashtag:
                    type: string
                    description: tag post
                    example: Sci-fi
        responses:
          '200':
            description: Successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Post'
          '400':
            description: Invalid request.
            content:
              application/json:
                examples:
                  Error:
                    description: Invalid request.
                    value:
                      message: Title and content cannot be empty!
          '401':
            description: Unauthorized author.
            content:
              application/json:
                examples:
                  Error:
                    description: Unauthorized author.
                    value:
                      message: Only author can update the post!
          '404':
            description: Not Found.
    /posts/movie:
      post:
        tags:
          - post
        summary: create a movie post
        description: create a movie post
        operationId: createMoviePost
        requestBody:
          description:
            create a movie post
          content:
            application/json:
              schema:
                type: object
                properties:
                  resourceId:
                    type: string
                    description: id of the external resource
                    example: xxxx
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Post'
          '400':
            description: 'resourceId cannot be empty'
          '404':
            description: Not found
...